created a MarketStatus component, separated Real time display, market status and market icon from Header component
